use float for rate

+++ SGPlayer/Classes/SGPlayerImp.h
@@ -100,7 +100,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)seekToTime:(NSTimeInterval)time;
 - (void)seekToTime:(NSTimeInterval)time completeHandler:(nullable void(^)(BOOL finished))completeHandler;
 
-@property (nonatomic, assign) CGFloat rate;
+@property (nonatomic, assign) float rate;
 
 @end
